fix: remove duplicate extension code
