--- app/src/main/java/eu/kanade/tachiyomi/extension/anime/api/AnimeExtensionGithubApi.kt
@@ -1,214 +1,0 @@
-package eu.kanade.tachiyomi.extension.anime.api
-
-import android.content.Context
-import ani.dantotsu.settings.saving.PrefManager
-import ani.dantotsu.settings.saving.PrefName
-import ani.dantotsu.util.Logger
-import eu.kanade.tachiyomi.extension.ExtensionUpdateNotifier
-import eu.kanade.tachiyomi.extension.anime.AnimeExtensionManager
-import eu.kanade.tachiyomi.extension.anime.model.AnimeExtension
-import eu.kanade.tachiyomi.extension.anime.model.AnimeLoadResult
-import eu.kanade.tachiyomi.extension.anime.model.AvailableAnimeSources
-import eu.kanade.tachiyomi.extension.util.ExtensionLoader
-import eu.kanade.tachiyomi.network.GET
-import eu.kanade.tachiyomi.network.NetworkHelper
-import eu.kanade.tachiyomi.network.awaitSuccess
-import eu.kanade.tachiyomi.network.parseAs
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.Json
-import tachiyomi.core.preference.Preference
-import tachiyomi.core.preference.PreferenceStore
-import tachiyomi.core.util.lang.withIOContext
-import uy.kohesive.injekt.injectLazy
-import java.util.Date
-import kotlin.time.Duration.Companion.days
-
-internal class AnimeExtensionGithubApi {
-
-    private val networkService: NetworkHelper by injectLazy()
-    private val preferenceStore: PreferenceStore by injectLazy()
-    private val animeExtensionManager: AnimeExtensionManager by injectLazy()
-    private val json: Json by injectLazy()
-
-    private val lastExtCheck: Preference<Long> by lazy {
-        preferenceStore.getLong("last_ext_check", 0)
-    }
-
-    suspend fun findExtensions(): List<AnimeExtension.Available> {
-        return withIOContext {
-
-            val extensions: ArrayList<AnimeExtension.Available> = arrayListOf()
-
-            val repos =
-                PrefManager.getVal<Set<String>>(PrefName.AnimeExtensionRepos).toMutableList()
-            if (repos.isEmpty()) {
-                repos.add("https://raw.githubusercontent.com/aniyomiorg/aniyomi-extensions/repo")
-                PrefManager.setVal(PrefName.AnimeExtensionRepos, repos.toSet())
-            }
-
-            repos.forEach {
-                try {
-                    val githubResponse = try {
-                        networkService.client
-                            .newCall(GET("${it}/index.min.json"))
-                            .awaitSuccess()
-                    } catch (e: Throwable) {
-                        Logger.log("Failed to get repo: $it")
-                        Logger.log(e)
-                        null
-                    }
-
-                    val response = githubResponse ?: run {
-                        networkService.client
-                            .newCall(GET(fallbackRepoUrl(it) + "/index.min.json"))
-                            .awaitSuccess()
-                    }
-
-                    val repoExtensions = with(json) {
-                        response
-                            .parseAs<List<AnimeExtensionJsonObject>>()
-                            .toExtensions(it)
-                    }
-
-                    // Sanity check - a small number of extensions probably means something broke
-                    // with the repo generator
-                    if (repoExtensions.size < 10) {
-                        throw Exception()
-                    }
-
-                    extensions.addAll(repoExtensions)
-                } catch (e: Throwable) {
-                    Logger.log("Failed to get extensions from GitHub")
-                    Logger.log(e)
-                }
-            }
-
-            extensions
-        }
-    }
-
-    suspend fun checkForUpdates(
-        context: Context,
-        fromAvailableExtensionList: Boolean = false
-    ): List<AnimeExtension.Installed>? {
-        // Limit checks to once a day at most
-        if (fromAvailableExtensionList && Date().time < lastExtCheck.get() + 1.days.inWholeMilliseconds) {
-            return null
-        }
-
-        val extensions = if (fromAvailableExtensionList) {
-            animeExtensionManager.availableExtensionsFlow.value
-        } else {
-            findExtensions().also { lastExtCheck.set(Date().time) }
-        }
-
-        val installedExtensions = ExtensionLoader.loadAnimeExtensions(context)
-            .filterIsInstance<AnimeLoadResult.Success>()
-            .map { it.extension }
-
-        val extensionsWithUpdate = mutableListOf<AnimeExtension.Installed>()
-        for (installedExt in installedExtensions) {
-            val pkgName = installedExt.pkgName
-            val availableExt = extensions.find { it.pkgName == pkgName } ?: continue
-
-            val hasUpdatedVer = availableExt.versionCode > installedExt.versionCode
-            val hasUpdatedLib = availableExt.libVersion > installedExt.libVersion
-            val hasUpdate = installedExt.isUnofficial.not() && (hasUpdatedVer || hasUpdatedLib)
-            if (hasUpdate) {
-                extensionsWithUpdate.add(installedExt)
-            }
-        }
-
-        if (extensionsWithUpdate.isNotEmpty()) {
-            ExtensionUpdateNotifier(context).promptUpdates(extensionsWithUpdate.map { it.name })
-        }
-
-        return extensionsWithUpdate
-    }
-
-    private fun List<AnimeExtensionJsonObject>.toExtensions(repository: String): List<AnimeExtension.Available> {
-        return this
-            .filter {
-                val libVersion = it.extractLibVersion()
-                libVersion >= ExtensionLoader.ANIME_LIB_VERSION_MIN && libVersion <= ExtensionLoader.ANIME_LIB_VERSION_MAX
-            }
-            .map {
-                AnimeExtension.Available(
-                    name = it.name.substringAfter("Aniyomi: "),
-                    pkgName = it.pkg,
-                    versionName = it.version,
-                    versionCode = it.code,
-                    libVersion = it.extractLibVersion(),
-                    lang = it.lang,
-                    isNsfw = it.nsfw == 1,
-                    hasReadme = it.hasReadme == 1,
-                    hasChangelog = it.hasChangelog == 1,
-                    sources = it.sources?.toAnimeExtensionSources().orEmpty(),
-                    apkName = it.apk,
-                    repository = repository,
-                    iconUrl = "${repository}/icon/${it.pkg}.png",
-                )
-            }
-    }
-
-    private fun List<AnimeExtensionSourceJsonObject>.toAnimeExtensionSources(): List<AvailableAnimeSources> {
-        return this.map {
-            AvailableAnimeSources(
-                id = it.id,
-                lang = it.lang,
-                name = it.name,
-                baseUrl = it.baseUrl,
-            )
-        }
-    }
-
-    fun getApkUrl(extension: AnimeExtension.Available): String {
-        return "${extension.repository}/apk/${extension.apkName}"
-    }
-
-    private fun fallbackRepoUrl(repoUrl: String): String? {
-        var fallbackRepoUrl = "https://gcore.jsdelivr.net/gh/"
-        val strippedRepoUrl =
-            repoUrl.removePrefix("https://").removePrefix("http://").removeSuffix("/")
-        val repoUrlParts = strippedRepoUrl.split("/")
-        if (repoUrlParts.size < 3) {
-            return null
-        }
-        val repoOwner = repoUrlParts[1]
-        val repoName = repoUrlParts[2]
-        fallbackRepoUrl += "$repoOwner/$repoName"
-        val repoBranch = if (repoUrlParts.size > 3) {
-            repoUrlParts[3]
-        } else {
-            "main"
-        }
-        fallbackRepoUrl += "@$repoBranch"
-        return fallbackRepoUrl
-    }
-}
-
-private fun AnimeExtensionJsonObject.extractLibVersion(): Double {
-    return version.substringBeforeLast('.').toDouble()
-}
-
-@Serializable
-private data class AnimeExtensionJsonObject(
-    val name: String,
-    val pkg: String,
-    val apk: String,
-    val lang: String,
-    val code: Long,
-    val version: String,
-    val nsfw: Int,
-    val hasReadme: Int = 0,
-    val hasChangelog: Int = 0,
-    val sources: List<AnimeExtensionSourceJsonObject>?,
-)
-
-@Serializable
-private data class AnimeExtensionSourceJsonObject(
-    val id: Long,
-    val lang: String,
-    val name: String,
-    val baseUrl: String,
-)
