+++ app/src/main/java/eu/kanade/tachiyomi/extension/api/ExtensionGithubApi.kt
@@ -1,14 +1,12 @@
-package eu.kanade.tachiyomi.extension.manga.api
+package eu.kanade.tachiyomi.extension.api
 
-import android.content.Context
 import ani.dantotsu.settings.saving.PrefManager
 import ani.dantotsu.settings.saving.PrefName
 import ani.dantotsu.util.Logger
-import eu.kanade.tachiyomi.extension.ExtensionUpdateNotifier
-import eu.kanade.tachiyomi.extension.manga.MangaExtensionManager
+import eu.kanade.tachiyomi.extension.anime.model.AnimeExtension
+import eu.kanade.tachiyomi.extension.anime.model.AvailableAnimeSources
 import eu.kanade.tachiyomi.extension.manga.model.AvailableMangaSources
 import eu.kanade.tachiyomi.extension.manga.model.MangaExtension
-import eu.kanade.tachiyomi.extension.manga.model.MangaLoadResult
 import eu.kanade.tachiyomi.extension.util.ExtensionLoader
 import eu.kanade.tachiyomi.network.GET
 import eu.kanade.tachiyomi.network.NetworkHelper
@@ -16,25 +14,143 @@ import eu.kanade.tachiyomi.network.awaitSuccess
 import eu.kanade.tachiyomi.network.parseAs
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
-import tachiyomi.core.preference.Preference
-import tachiyomi.core.preference.PreferenceStore
 import tachiyomi.core.util.lang.withIOContext
 import uy.kohesive.injekt.injectLazy
-import java.util.Date
-import kotlin.time.Duration.Companion.days
 
-internal class MangaExtensionGithubApi {
-
+internal class ExtensionGithubApi {
     private val networkService: NetworkHelper by injectLazy()
-    private val preferenceStore: PreferenceStore by injectLazy()
-    private val extensionManager: MangaExtensionManager by injectLazy()
     private val json: Json by injectLazy()
 
-    private val lastExtCheck: Preference<Long> by lazy {
-        preferenceStore.getLong("last_ext_check", 0)
+    private fun List<ExtensionSourceJsonObject>.toAnimeExtensionSources(): List<AvailableAnimeSources> {
+        return this.map {
+            AvailableAnimeSources(
+                id = it.id,
+                lang = it.lang,
+                name = it.name,
+                baseUrl = it.baseUrl,
+            )
+        }
     }
 
-    suspend fun findExtensions(): List<MangaExtension.Available> {
+    private fun List<ExtensionJsonObject>.toAnimeExtensions(repository: String): List<AnimeExtension.Available> {
+        return this
+            .filter {
+                val libVersion = it.extractLibVersion()
+                libVersion >= ExtensionLoader.ANIME_LIB_VERSION_MIN && libVersion <= ExtensionLoader.ANIME_LIB_VERSION_MAX
+            }
+            .map {
+                AnimeExtension.Available(
+                    name = it.name.substringAfter("Aniyomi: "),
+                    pkgName = it.pkg,
+                    versionName = it.version,
+                    versionCode = it.code,
+                    libVersion = it.extractLibVersion(),
+                    lang = it.lang,
+                    isNsfw = it.nsfw == 1,
+                    hasReadme = it.hasReadme == 1,
+                    hasChangelog = it.hasChangelog == 1,
+                    sources = it.sources?.toAnimeExtensionSources().orEmpty(),
+                    apkName = it.apk,
+                    repository = repository,
+                    iconUrl = "${repository}/icon/${it.pkg}.png",
+                )
+            }
+    }
+
+    suspend fun findAnimeExtensions(): List<AnimeExtension.Available> {
+        return withIOContext {
+
+            val extensions: ArrayList<AnimeExtension.Available> = arrayListOf()
+
+            val repos =
+                PrefManager.getVal<Set<String>>(PrefName.MangaExtensionRepos).toMutableList()
+            if (repos.isEmpty()) {
+                repos.add("https://raw.githubusercontent.com/keiyoushi/extensions/main")
+                PrefManager.setVal(PrefName.MangaExtensionRepos, repos.toSet())
+            }
+
+            repos.forEach {
+                try {
+                    val githubResponse = try {
+                        networkService.client
+                            .newCall(GET("${it}/index.min.json"))
+                            .awaitSuccess()
+                    } catch (e: Throwable) {
+                        Logger.log("Failed to get repo: $it")
+                        Logger.log(e)
+                        null
+                    }
+
+                    val response = githubResponse ?: run {
+                        networkService.client
+                            .newCall(GET(fallbackRepoUrl(it) + "/index.min.json"))
+                            .awaitSuccess()
+                    }
+
+                    val repoExtensions = with(json) {
+                        response
+                            .parseAs<List<ExtensionJsonObject>>()
+                            .toAnimeExtensions(it)
+                    }
+
+                    // Sanity check - a small number of extensions probably means something broke
+                    // with the repo generator
+                    if (repoExtensions.size < 10) {
+                        throw Exception()
+                    }
+
+                    extensions.addAll(repoExtensions)
+                } catch (e: Throwable) {
+                    Logger.log("Failed to get extensions from GitHub")
+                    Logger.log(e)
+                }
+            }
+
+            extensions
+        }
+    }
+
+    fun getAnimeApkUrl(extension: AnimeExtension.Available): String {
+        return "${extension.repository}/apk/${extension.apkName}"
+    }
+
+    private fun List<ExtensionSourceJsonObject>.toMangaExtensionSources(): List<AvailableMangaSources> {
+        return this.map {
+            AvailableMangaSources(
+                id = it.id,
+                lang = it.lang,
+                name = it.name,
+                baseUrl = it.baseUrl,
+            )
+        }
+    }
+
+    private fun List<ExtensionJsonObject>.toMangaExtensions(repository: String): List<MangaExtension.Available> {
+        return this
+            .filter {
+                val libVersion = it.extractLibVersion()
+                libVersion >= ExtensionLoader.MANGA_LIB_VERSION_MIN && libVersion <= ExtensionLoader.MANGA_LIB_VERSION_MAX
+            }
+            .map {
+                MangaExtension.Available(
+                    name = it.name.substringAfter("Tachiyomi: "),
+                    pkgName = it.pkg,
+                    versionName = it.version,
+                    versionCode = it.code,
+                    libVersion = it.extractLibVersion(),
+                    lang = it.lang,
+                    isNsfw = it.nsfw == 1,
+                    hasReadme = it.hasReadme == 1,
+                    hasChangelog = it.hasChangelog == 1,
+                    sources = it.sources?.toMangaExtensionSources().orEmpty(),
+                    apkName = it.apk,
+                    repository = repository,
+                    iconUrl = "${repository}/icon/${it.pkg}.png",
+                )
+            }
+    }
+
+    suspend fun findMangaExtensions(): List<MangaExtension.Available> {
         return withIOContext {
 
             val extensions: ArrayList<MangaExtension.Available> = arrayListOf()
@@ -67,7 +183,7 @@ internal class MangaExtensionGithubApi {
                     val repoExtensions = with(json) {
                         response
                             .parseAs<List<ExtensionJsonObject>>()
-                            .toExtensions(it)
+                            .toMangaExtensions(it)
                     }
 
                     // Sanity check - a small number of extensions probably means something broke
@@ -87,86 +203,8 @@ internal class MangaExtensionGithubApi {
         }
     }
 
-    suspend fun checkForUpdates(
-        context: Context,
-        fromAvailableExtensionList: Boolean = false
-    ): List<MangaExtension.Installed>? {
-        // Limit checks to once a day at most
-        if (fromAvailableExtensionList && Date().time < lastExtCheck.get() + 1.days.inWholeMilliseconds) {
-            return null
-        }
-
-        val extensions = if (fromAvailableExtensionList) {
-            extensionManager.availableExtensionsFlow.value
-        } else {
-            findExtensions().also { lastExtCheck.set(Date().time) }
-        }
-
-        val installedExtensions = ExtensionLoader.loadMangaExtensions(context)
-            .filterIsInstance<MangaLoadResult.Success>()
-            .map { it.extension }
-
-        val extensionsWithUpdate = mutableListOf<MangaExtension.Installed>()
-        for (installedExt in installedExtensions) {
-            val pkgName = installedExt.pkgName
-            val availableExt = extensions.find { it.pkgName == pkgName } ?: continue
-            val hasUpdatedVer = availableExt.versionCode > installedExt.versionCode
-            val hasUpdatedLib = availableExt.libVersion > installedExt.libVersion
-            val hasUpdate = installedExt.isUnofficial.not() && (hasUpdatedVer || hasUpdatedLib)
-            if (hasUpdate) {
-                extensionsWithUpdate.add(installedExt)
-            }
-        }
-
-        if (extensionsWithUpdate.isNotEmpty()) {
-            ExtensionUpdateNotifier(context).promptUpdates(extensionsWithUpdate.map { it.name })
-        }
-
-        return extensionsWithUpdate
-    }
-
-    private fun List<ExtensionJsonObject>.toExtensions(repository: String): List<MangaExtension.Available> {
-        return this
-            .filter {
-                val libVersion = it.extractLibVersion()
-                libVersion >= ExtensionLoader.MANGA_LIB_VERSION_MIN && libVersion <= ExtensionLoader.MANGA_LIB_VERSION_MAX
-            }
-            .map {
-                MangaExtension.Available(
-                    name = it.name.substringAfter("Tachiyomi: "),
-                    pkgName = it.pkg,
-                    versionName = it.version,
-                    versionCode = it.code,
-                    libVersion = it.extractLibVersion(),
-                    lang = it.lang,
-                    isNsfw = it.nsfw == 1,
-                    hasReadme = it.hasReadme == 1,
-                    hasChangelog = it.hasChangelog == 1,
-                    sources = it.sources?.toExtensionSources().orEmpty(),
-                    apkName = it.apk,
-                    repository = repository,
-                    iconUrl = "${repository}/icon/${it.pkg}.png",
-                )
-            }
-    }
-
-    private fun List<ExtensionSourceJsonObject>.toExtensionSources(): List<AvailableMangaSources> {
-        return this.map {
-            AvailableMangaSources(
-                id = it.id,
-                lang = it.lang,
-                name = it.name,
-                baseUrl = it.baseUrl,
-            )
-        }
-    }
-
-    fun getApkUrl(extension: MangaExtension.Available): String {
+    fun getMangaApkUrl(extension: MangaExtension.Available): String {
         return "${extension.repository}/apk/${extension.apkName}"
-    }
-
-    private fun ExtensionJsonObject.extractLibVersion(): Double {
-        return version.substringBeforeLast('.').toDouble()
     }
 
     private fun fallbackRepoUrl(repoUrl: String): String? {
@@ -211,3 +249,7 @@ private data class ExtensionSourceJsonObject(
     val name: String,
     val baseUrl: String,
 )
+
+private fun ExtensionJsonObject.extractLibVersion(): Double {
+    return version.substringBeforeLast('.').toDouble()
+}
